Move hosting extensions to Hosting folder
